requirements.txt
# Twitter API Client
tweepy>=4.14.0

# AI/ML - Anthropic Claude
anthropic>=0.18.0

# HTTP Requests
requests>=2.31.0

# Environment Configuration
python-dotenv>=1.0.0

# YAML Configuration
pyyaml>=6.0.1

# Slack Integration
slack-sdk>=3.27.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
